Fix path resolution to checks
We can't assume a directory of `./src/checker/checks` for our checks
when `sql-lint` has been packaged into an executable. `pkg` has its own
snapshot filesystem which we use. In this case we change it to the more
robust `__dirname` and everything works.

Along with this, we have slightly refactored the way that checks are NOT
included into the final checking. Whereas before it was based on index
of the file in the list (very fragile), it's now based on name and
extension.