Remove postinstall commands
The postinstall command is causing issue for Windows users. I was
originally going to move it over to a `postinstall.js` file and change
the postinstall to `node build/postinstall.js` to make it more portable
but I think overall, the postinstall is pointless. There are already
many warnings that you should use a configuration file for a better
experience.

See here:
https://github.com/joereynolds/sql-lint/issues/157
and
https://github.com/github/super-linter/issues/538