No longer crash on uncategoriseable query
This used to be useful when sql-lint was first starting off but now it
just gets in the way. Instead, we log out to the console if we can't
categorise a query. Note that this warning will not show in the
integration with vim. To see these errors you'll need to manually run
`sql-lint` (which I imagine most people do) like this:

```
sql-lint your-file.sql
```