fix: Handle multiline sql without spaces
When replacing multiline comment, it is now done with a space instead of
nothing since this would be valid sql:
```sql
SELECT pid FROM/*comment*/pg_stat_activity;
```
But replacing the comment with "" would make it incorrect.

Also join the lines in query on getContent with a space to remove the
need for spaces in the sql. Previously this sql:
```sql
SELECT some_column
FROM some_table;
```
would result in this error: `[error] column "some_columnfrom" does not
exist`
This PR fixes that.