Fix ignore-errors not being respected
Even though this was a config option, it was never actually honoured for
some reason. Currently this option can only be done via config.json but
it'd be good if we could specify it as a command line option too i.e.
```
sql-lint --ignore-errors=missing-where,trailing-whitespace
```

This would also allow us to write some integration tests against it
because currently there aren't any tests against this :(