Feature/allow rename table syntax (#222)
* Delete unused src/lexer/keyword.ts

* Add support for `RENAME` statement

This adds in support for the `RENAME` since this was breaking some
statements. Alongside this we have also lexed/tokenised it for the most
part. We make the assumption that anything that isn't "RENAME" or "TO"
is a `table_reference`.

Note: We have only done this for mysql for now. I imagine it can be copy
pasted into the other dialects (assuming there's no deviation in how
they work)

* Trigger build

Co-authored-by: Joe Reynolds <joe.reynolds@sykescottages.co.uk>